requirements.txt
-i https://pypi.org/simple
aniso8601==6.0.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
dialogflow==0.5.2
flask-assistant==0.3.91
flask-cors==3.0.7
flask==1.0.2
google-api-core[grpc]==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-logging==1.10.0
googleapis-common-protos[grpc]==1.5.9
grpcio==1.19.0
gunicorn==19.9.0
honcho==1.0.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10
markupsafe==1.1.1
protobuf==3.7.0
pyasn1-modules==0.2.4
pyasn1==0.4.5
pytz==2018.9
requests==2.21.0
rsa==4.0
ruamel.yaml==0.15.89
six==1.12.0
urllib3==1.24.1
werkzeug==0.15.1
